optimize trip command related operation
