+++ src/main/java/seedu/address/commons/core/Messages.java
@@ -9,6 +9,7 @@ public class Messages {
     public static final String MESSAGE_INVALID_COMMAND_FORMAT = "Invalid command format! \n%1$s";
     public static final String MESSAGE_INVALID_PASSENGER_DISPLAYED_INDEX = "The passenger index provided is invalid";
     public static final String MESSAGE_PASSENGER_LISTED_OVERVIEW = "%1$d passengers listed!";
+    public static final String MESSAGE_TRIPS_LISTED_OVERVIEW = "%1$d trips listed!";
 
     /**
      * Prevents Messages from being instantiated.
